pyproject.toml
[tool.poetry]
name = "lista-de-tarefas"
version = "0.1.0"
description = ""
authors = ["Thi <tiagoqueirozgoncalves100@gmail.com>"]
readme = "README.md"
packages = [{include = "lista_de_tarefas"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
python-decouple = "^3.6"
Flask-WTF = "^1.0.1"
Flask-Login = "^0.6.2"
flask-sqlalchemy = "^3.0.2"
Flask-Migrate = "^3.1.0"
email-validator = "^1.3.0"
Flask-Babel = "^2.0.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.1"
PyMySQL = "^1.0.2"
cryptography = "^38.0.1"
pyjwt = "^2.6.0"
flasgger = "^0.9.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
